fix: API校验hours字段去掉.int()限制,允许小数课时
validation.js中paymentSchema和teacherSchema的hours字段
原来是z.number().int().min(0), 6.6是小数校验失败返回400
改为z.number().min(0), 允许6.6等小数课时通过校验

+++ api/src/utils/validation.js
@@ -109,7 +109,7 @@ export const paymentSchema = z.object({
   date: z.string().optional(),
   receipt_number: z.string().max(50).optional().or(z.literal('')),
   notes: z.string().optional().or(z.literal('')),
-  hours: z.number().int().min(0).optional().default(0)
+  hours: z.number().min(0).optional().default(0)
 });
 
 // Teachers Schema
@@ -121,7 +121,7 @@ export const teacherSchema = z.object({
   hourly_rate: z.number().min(0).optional(),
   status: z.enum(['active', 'inactive']).optional().default('active'),
   notes: z.string().optional().or(z.literal('')),
-  hours: z.number().int().min(0).optional().default(0)
+  hours: z.number().min(0).optional().default(0)
 });
 
 export const teacherUpdateSchema = teacherSchema.partial();
